refactor(react-gui): close the thread boundary and enforce it as an error
The renderer <-> worker layering was declared but only warned about, and two
crossings had already shipped as real bugs: a File Open pane imported
`probeMapHeader.service` at RUNTIME, pulling worker code that expects `fs`
into the renderer bundle, and the colouring service imported
`components/multigrad/multiGradPresets`, pulling a components module into the
worker bundle. Nine more services reached up into the dialog tree for their
own argument types.

Every shape both threads name now lives on the boundary:

  worker/shared/fileOpenTypes.ts     what the File Open dialog decided
                                     (the helpers that PRODUCE it stay in
                                     components/fopen-opt-dlgs/types.ts)
  worker/shared/mapHeader.ts         map header facts + the subsample rule
  worker/shared/multiGradPresets.ts  gradient presets
  worker/shared/eventConst.ts        ScrEventManager filter constants

The old modules re-export them, so no UI file changes. The last 30 warnings
were deep relative imports climbing past the alias threshold; those are now
`@renderer/` paths.

With the tree clean the boundary rules move from `warn` to `error` -- a
warning would not have stopped either bug. Verified in the built bundles: the
worker bundle contains no Blueprint, React or components/ code, and the
renderer bundle no worker service internals.
