+++ tritium/react-gui/src/renderer/components/fopen-opt-dlgs/types.ts
@@ -12,107 +12,34 @@
  */
 import type { ReaderDefaultOptions } from '../../worker/server/services/getReaderDefaultOptions.service';
 
-// ---- Format kind ----
+// The option types themselves live on the worker boundary (nine services read
+// them); re-exported here so dialog code can keep taking them from one place.
+export type {
+    FormatKind,
+    PdbOptions,
+    MtzOptions,
+    MapTypeChoice,
+    Ccp4MapOptions,
+    MsmsOptions,
+    NamdCoorOptions,
+    AmberPrmtopOptions,
+    RendererOptions,
+    PresetTypeEntry,
+    FormatOptions,
+    FileOpenOptions,
+} from '../../worker/shared/fileOpenTypes';
 
-export type FormatKind = 'pdb' | 'mmcif' | 'mtz' | 'ccp4map' | 'msms' | 'namdcoor' | 'amberprm' | 'unknown';
-
-// ---- Per-format option types ----
-
-export interface PdbOptions {
-  loadModel: boolean;
-  loadAnisou: boolean;
-  loadAltConf: boolean;
-  loadSegid: boolean;
-  build2ndry: boolean;
-  autoTopology: boolean;
-}
-
-export interface MtzOptions {
-  columnF: string;
-  columnPhi: string;
-  /** Whether the phase column is used (UXP "Phase" checkbox). */
-  phaseEnabled: boolean;
-  columnW: string;
-  /** Whether the weight column is used (UXP "Weight" checkbox). */
-  weightEnabled: boolean;
-  resolutionLimit: number;
-  gridSpacing: number;
-}
-
-/** DensityMap `map_type` values offered by the CCP4/MRC option pane. */
-export type MapTypeChoice = 'auto' | 'xtal' | 'em';
-
-export interface Ccp4MapOptions {
-  normalize: boolean;
-  truncateMinEnabled: boolean;
-  truncateMin: number;
-  truncateMaxEnabled: boolean;
-  truncateMax: number;
-  /**
-   * Map kind override applied to the loaded DensityMap (`map_type`), not a
-   * reader property: 'auto' keeps the header-based detection.
-   */
-  mapType: MapTypeChoice;
-  /** CCP4MapReader `subsample`: keep every n-th grid point on each axis. */
-  subsample: number;
-}
-
-export interface MsmsOptions {
-  vertFilePath: string;
-}
-
-export interface NamdCoorOptions {
-  psfFilePath: string;
-}
-
-export interface AmberPrmtopOptions {
-  /**
-   * Optional AMBER coordinate file (inpcrd / rst7 / restrt) attached as the
-   * reader's "coord" sub-stream. Empty -> topology-only load (atoms at zero).
-   */
-  coordFilePath: string;
-}
-
-export interface RendererOptions {
-  objectName: string;
-  rendererType: string;
-  rendererName: string;
-  selectionEnabled: boolean;
-  selection: string;
-  centerView: boolean;
-  /**
-   * Set when the user picked a renderer preset (a `<objtype>-rendpreset`
-   * style name, e.g. 'Default1RendPreset'). When set, the worker creates a
-   * renderer group + child renderers via `Object.createPresetRenderer` and
-   * `rendererType` is ignored; `rendererName` becomes both the group name
-   * and the child-name prefix. Undefined -> plain single-renderer create.
-   */
-  presetName?: string;
-}
-
-/** One preset entry offered in the renderer-type dropdown. */
-export interface PresetTypeEntry {
-  /** Style id, e.g. 'Default1RendPreset' (the value passed to the worker). */
-  name: string;
-  /** Human label from the style's desc attribute ('' when undefined). */
-  desc: string;
-}
-
-// Discriminated union for format-specific options
-export type FormatOptions =
-  | { kind: 'pdb'; options: PdbOptions }
-  | { kind: 'mmcif'; options: PdbOptions }
-  | { kind: 'mtz'; options: MtzOptions }
-  | { kind: 'ccp4map'; options: Ccp4MapOptions }
-  | { kind: 'msms'; options: MsmsOptions }
-  | { kind: 'namdcoor'; options: NamdCoorOptions }
-  | { kind: 'amberprm'; options: AmberPrmtopOptions }
-  | { kind: 'unknown'; options: Record<string, never> };
-
-export interface FileOpenOptions {
-  format: FormatOptions;
-  renderer: RendererOptions;
-}
+import type {
+    FormatKind,
+    PdbOptions,
+    MtzOptions,
+    Ccp4MapOptions,
+    MsmsOptions,
+    NamdCoorOptions,
+    AmberPrmtopOptions,
+    RendererOptions,
+    FormatOptions,
+} from '../../worker/shared/fileOpenTypes';
 
 // ---- Format detection ----
