+++ tritium/react-gui/src/renderer/components/panels/mdtraj/TrajTrack.tsx
@@ -12,7 +12,7 @@
  */
 
 import React, { useCallback, useRef, useState } from 'react';
-import type { TrajBlockInfo } from '../../../worker/server/services/trajectory.service';
+import type { TrajBlockInfo } from '@renderer/worker/server/services/trajectory.service';
 import { TrajBlockStrip } from './TrajBlockStrip';
 import {
     frameToPx,
