+++ tritium/react-gui/src/renderer/components/panes/settings/SettingRow.tsx
@@ -15,10 +15,10 @@ import {
   ColorField,
   TextField,
   FormButton,
-} from '../../../h3-kit/form'
+} from '@renderer/h3-kit/form'
 import { IPC } from '@shared/ipcChannels'
 import type { SettingDef } from './settingsConfig'
-import type { Mode } from '../../../h3-kit/colorpicker/ColorPicker'
+import type { Mode } from '@renderer/h3-kit/colorpicker/ColorPicker'
 
 /**
  * App settings colours are scene-independent plain colours, so the picker
