+++ tritium/react-gui/src/renderer/event.ts
@@ -1,21 +1,10 @@
+/**
+ * @file renderer/event.ts
+ * @description CueMol event-manager filter constants, re-exported.
+ *
+ * The definitions live in `worker/shared/eventConst.ts` because the worker
+ * names them too; this module is the renderer-side spelling that the hooks
+ * import.
+ */
 
-export const SEM_ANY = -1
-
-// source category ID definition
-export const SEM_LOG = 0x0001
-export const SEM_INDEV = 0x0002
-export const SEM_SCENE = 0x0004
-export const SEM_OBJECT = 0x0008
-export const SEM_RENDERER = 0x0010
-export const SEM_VIEW = 0x0020
-export const SEM_CAMERA = 0x0040
-export const SEM_STYLE = 0x0080
-export const SEM_ANIM = 0x0100
-export const SEM_EXTND = 0x8000
-
-// event type ID definition
-export const SEM_ADDED = 1
-export const SEM_REMOVING = 2
-export const SEM_PROPCHG = 3
-export const SEM_CHANGED = 4
-export const SEM_OTHER = 9999
+export * from './worker/shared/eventConst';
