+++ tritium/react-gui/src/renderer/worker/server/services/bondEdit.service.ts
@@ -29,7 +29,7 @@ import type { MsgLog } from '@cuemol/core/src/wrappers/MsgLog';
 import type { DistPickDrawObj } from '@cuemol/core/src/wrappers/DistPickDrawObj';
 import type { MolCoord } from '@cuemol/core/src/wrappers/MolCoord';
 import type { MolAnlManager } from '@cuemol/core/src/wrappers/MolAnlManager';
-import type { HitTestResult } from '../../../types';
+import type { HitTestResult } from '@renderer/types';
 import { withUndoTxn } from './withUndoTxn';
 import { getSceneOrNull } from './helpers/sceneResolver';
