+++ tritium/react-gui/src/renderer/worker/server/services/renderBackends/PovrayBackend.ts
@@ -17,8 +17,8 @@ import type { AnimMgr } from "@cuemol/core/src/wrappers/AnimMgr";
 import type { Scene } from "@cuemol/core/src/wrappers/Scene";
 import type { PovSceneExporter } from "@cuemol/core/src/wrappers/PovSceneExporter";
 import type { WorkerContext } from "../../types/WorkerContext";
-import type { RenderSettingsSnapshot } from "../../../../data/renderResult";
-import type { RenderBinaries } from "../../../shared/renderTypes";
+import type { RenderSettingsSnapshot } from "@renderer/data/renderResult";
+import type { RenderBinaries } from "@renderer/worker/shared/renderTypes";
 import { quoteProcArg } from "./procArgs";
 import {
   type RenderBackend,
