+++ tritium/react-gui/src/renderer/worker/shared/eventConst.ts
@@ -1,0 +1,31 @@
+/**
+ * @file worker/shared/eventConst.ts
+ * @description `qlib::ScrEventManager` filter constants.
+ *
+ * Both threads name these: the renderer to subscribe, the worker to register
+ * its own listeners. They live on the boundary so `worker/server/` does not
+ * have to import a renderer module (`renderer/event.ts`, which re-exports
+ * them for the 30-odd hooks that have always taken them from there).
+ */
+
+
+export const SEM_ANY = -1
+
+// source category ID definition
+export const SEM_LOG = 0x0001
+export const SEM_INDEV = 0x0002
+export const SEM_SCENE = 0x0004
+export const SEM_OBJECT = 0x0008
+export const SEM_RENDERER = 0x0010
+export const SEM_VIEW = 0x0020
+export const SEM_CAMERA = 0x0040
+export const SEM_STYLE = 0x0080
+export const SEM_ANIM = 0x0100
+export const SEM_EXTND = 0x8000
+
+// event type ID definition
+export const SEM_ADDED = 1
+export const SEM_REMOVING = 2
+export const SEM_PROPCHG = 3
+export const SEM_CHANGED = 4
+export const SEM_OTHER = 9999
